pyproject.toml
[tool.poetry]
name = "lambdanator"
version = "0.1.0"
description = "combination of lambda bots"
authors = ["StarrFox <starrfox6312@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.6.0"
humanize = "^3.3.0"
toml = "^0.10.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
